tests/wpt/web-platform-tests/webmessaging/with-ports/025.html

<!doctype html>
<title>1 as ports</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
test(function() {
   assert_throws(new TypeError(), function() {
      postMessage('', '*', 1);
   });
});
</script>
